requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.0
pydantic==2.4.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
google-generativeai==0.3.0